chore: extract verifyModelWithConfig and prepareModelCOnfig to configState
